# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 4 */
  password: string;
  displayName?: string;
}

export interface User {
  id: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  balance: number;
  isAdmin: boolean;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface BalanceResponse {
  balance: number;
}

export interface DepositInput {
  /** @minimum 1000 */
  amount: number;
}

export interface BetInput {
  choice: string;
  /** @minimum 1000 */
  amount: number;
}

export interface BetRecord {
  id: number;
  choice: string;
  amount: number;
  won: boolean;
  payout: number;
  /** @nullable */
  dice1?: number | null;
  /** @nullable */
  dice2?: number | null;
  /** @nullable */
  dice3?: number | null;
  createdAt: string;
}

export interface BetResult {
  won: boolean;
  payout: number;
  dice1: number;
  dice2: number;
  dice3: number;
  balance: number;
}

export interface AdminUser {
  id: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  balance: number;
  isAdmin: boolean;
  totalBets: number;
  createdAt: string;
}

export interface AdjustBalanceInput {
  amount: number;
  note: string;
}

export interface ResetPasswordInput {
  /** @minLength 4 */
  newPassword: string;
}

export interface LeaderboardEntry {
  username: string;
  /** @nullable */
  displayName?: string | null;
  totalBets: number;
  totalWon: number;
  winRate: number;
}

export interface GameStats {
  totalBets: number;
  totalWon: number;
  winRate: number;
  biggestWin: number;
}