# lib/api-zod/src/generated/types/betRecord.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BetRecord {
  id: number;
  choice: string;
  amount: number;
  won: boolean;
  payout: number;
  /** @nullable */
  dice1?: number | null;
  /** @nullable */
  dice2?: number | null;
  /** @nullable */
  dice3?: number | null;
  createdAt: string;
}