# lib/api-zod/src/generated/types/betResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BetResult {
  won: boolean;
  payout: number;
  dice1: number;
  dice2: number;
  dice3: number;
  balance: number;
}