# lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  username: string;
  /** @nullable */
  displayName?: string | null;
  totalBets: number;
  totalWon: number;
  winRate: number;
}