# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: number;
  username: string;
  /** @nullable */
  displayName?: string | null;
  balance: number;
  isAdmin: boolean;
}